Feat 0.1.12/recent memorial/backend 43

+++ src/main/java/windeath44/server/memorial/domain/service/MemorialCommentLikesService.java
@@ -4,14 +4,13 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import windeath44.server.memorial.domain.mapper.MemorialCommentLikesMapper;
-import windeath44.server.memorial.domain.model.MemorialCommentLikesCount;
 import windeath44.server.memorial.domain.model.MemorialComment;
 import windeath44.server.memorial.domain.model.MemorialCommentLikes;
 import windeath44.server.memorial.domain.model.MemorialCommentLikesPrimaryKey;
-import windeath44.server.memorial.domain.model.vo.MemorialCommentLikesCountList;
 import windeath44.server.memorial.domain.repository.MemorialCommentLikesRepository;
 
-import java.util.List;
+import java.util.Collections;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -23,21 +22,30 @@ public class MemorialCommentLikesService {
   @Transactional
   public void like(Long commentId, String userId) {
     MemorialComment comment = memorialCommentService.findCommentById(commentId);
+
     MemorialCommentLikesPrimaryKey memorialCommentLikesPrimaryKey = comment.likesKey(userId);
+    if (memorialCommentLikesRepository.existsById(memorialCommentLikesPrimaryKey)) return;
+
     MemorialCommentLikes memorialCommentLikes = memorialCommentLikesMapper.toMemorialCommentLikes(memorialCommentLikesPrimaryKey);
     memorialCommentLikesRepository.save(memorialCommentLikes);
+    comment.upLikes();
   }
 
   @Transactional
   public void unlike(Long commentId, String userId) {
     MemorialComment comment = memorialCommentService.findCommentById(commentId);
+
     MemorialCommentLikesPrimaryKey memorialCommentLikesPrimaryKey = comment.likesKey(userId);
+    if (!memorialCommentLikesRepository.existsById(memorialCommentLikesPrimaryKey)) return;
+
     memorialCommentLikesRepository.deleteById(memorialCommentLikesPrimaryKey);
+    comment.downLikes();
   }
 
-  public MemorialCommentLikesCountList getLikesCountByCommentIds(List<Long> commentIds) {
-    List<MemorialCommentLikesCount> memorialCommentLikesCounts = memorialCommentLikesRepository.findLikesCountGroupedByCommentIds(commentIds);
-    return MemorialCommentLikesCountList.of(memorialCommentLikesCounts);
+  public Set<Long> getLikedCommentIds(String userId, Set<Long> commentIds) {
+    if (commentIds == null || commentIds.isEmpty()) {
+      return Collections.emptySet();
+    }
+    return memorialCommentLikesRepository.findLikedCommentIdsByUserIdAndCommentIds(userId, commentIds);
   }
-
 }
